backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
clang==6.0.0.2
python-multipart==0.0.6
aiofiles==23.2.1
tree-sitter==0.20.4
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pyjsparser==2.7.1
bcrypt==3.2.0
semgrep==1.60.0 